// src/Admin/Models/AmbassdorDetails.ts
export default class AmbassdorDetails {
  id: number = -1;
  first_name: string = "";
  last_name: string = "";
  email: string = "";
  noOfProgram: string = "";
  Programid: string = "";
  ProgramNames: string = "";
  other_details: string = "";
  url: string = "";
  about: string = "";
  industry: string = "";
  country: string = "";
  status: boolean = true;
  ambassador_img: string = "";
  program_img: string = "";
}